Lint: pre-commit
